requirements.txt
aioredis==1.0.0
appnope==0.1.0
asgiref==2.2.0
astroid==1.6.3
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
channels==2.0.2
channels-redis==2.1.1
colorama==0.3.9
constantly==15.1.0
daphne==2.1.0
decorator==4.2.1
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.7.7
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.11.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.5.6
mysqlclient==1.3.12
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.8.4
pytz==2018.3
selenium==3.11.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
txaio==2.9.0
wcwidth==0.1.7
wrapt==1.10.11
zope.interface==4.4.3
